+++ src/Command/Install/CreateVendorsLinksCommand.php
@@ -50,8 +50,8 @@ class CreateVendorsLinksCommand extends Command
         foreach (Configure::readFromPlugins('VendorLinks') as $origin => $target) {
             $this->createLink(
                 $io,
-                $Filesystem->concatenate(ROOT, 'vendor', $Filesystem->normalizePath($origin)),
-                $Filesystem->concatenate(WWW_ROOT, 'vendor', $target)
+                VENDOR . $Filesystem->normalizePath($origin),
+                WWW_VENDOR . $target
             );
         }
     }